requirements.txt
google-cloud-aiplatform
langchain
streamlit
IPython